pyproject.toml
[project]
name = "hands-on-recommender-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "~=3.11"
dependencies = [
    "altair>=4.2.2",
    "catboost==1.2",
    "huggingface-hub==0.24.7",
    "ipykernel>=6.29.5",
    "langchain-openai==0.1.14",
    "langchain==0.2.6",
    "loguru>=0.7.2",
    "nbformat>=5.10.4",
    "polars==1.9.0",
    "pydantic-settings>=2.6.1",
    "sentence-transformers==2.2.2",
    "streamlit==1.28.2",
    "tensorflow-recommenders==0.7.2",
    "tensorflow==2.14",
    "pyproject-toml>=0.0.12",
    "pandas-stubs>=2.2.3.250308",
    "mlflow>=2.21.1",
    "types-tqdm>=4.67.0.20250319",
]

[dependency-groups]
dev = [
    "ruff>=0.7.2",
]
